require feature 'state' (introduced in Perl 5.10.0)
The META spec has no way of specifying "I want a Perl having 'state'
variables", so set MIN_PERL_VERSION n Makefile.PL to the version that
introduced this feature. But "use feature 'state'" instead of "use 5.10.0"
in the modules as the latter also brings in crap introduced in 5.10,
e.g. "switch".