:tied_objects: silently omit "source" from the keys of the tied hash
You can't "get" the "source" attribute, hence ignore it when
iterating over the hash. This gets rid of the WTF moment, when
something inadvertently tries to enumerate the contents of the tied
hash. Examples include calling Data::Dumper::Dumper() on a
VCS::CMSynergy::Object (e.g. Devel::ptkdb does this when you "mouse over"
such a value) or Type::Params trying to create a description of
a value that violates the parameter constraint for its error message.